requirements.txt
Pillow==9.1.1
PyMySQL==1.0.2
